+++ core/constants.py
@@ -41,8 +41,6 @@ TEMPLATE_MAPPING = {
     # Invest
     'InvestInternationalHomePage': 'invest/landing_page.html',
     'InvestHighPotentialOpportunityDetailPage': 'invest/hpo/high_potential_opportunity_detail.html',
-    'InvestHighPotentialOpportunityFormPage': 'invest/hpo/high_potential_opportunities_form.html',
-    'InvestHighPotentialOpportunityFormSuccessPage': 'invest/hpo/high_potential_opportunities_form_success.html',
     'InvestRegionPage': 'invest/regions/region_detail.html',
 
     # Find a supplier
@@ -56,6 +54,10 @@ TEMPLATE_MAPPING = {
     'InternationalInvestmentSubSectorPage': 'investment_atlas/sector.html',
     'InvestmentGeneralContentPage': 'investment_atlas/general_content_page.html',
     'WhyInvestInTheUKPage': 'investment_atlas/why_invest_in_the_uk_page.html',
+    'InvestForeignDirectInvestmentOpportunityFormPage':
+        'investment_atlas/fdi_opportunities_form.html',  # Not used? has custom view
+    'InvestForeignDirectInvestmentOpportunityFormSuccessPage':
+        'investment_atlas/fdi_opportunities_form_success.html',  # Not used? has custom view
 }
 
 FEATURE_FLAGGED_URLS_MAPPING = {
@@ -176,14 +178,14 @@ GA_DATA_MAPPING = {
         'site_section': 'HighPotentialOpportunities',
         'site_subsection': 'DetailPage',
     },
-    'InvestHighPotentialOpportunityFormPage': {
+    'ForeignDirectInvestmentFormPage': {
         'business_unit': 'Invest',
-        'site_section': 'HighPotentialOpportunities',
+        'site_section': 'ForeignDirectInvestment',
         'site_subsection': 'FormPage',
     },
-    'InvestHighPotentialOpportunityFormSuccessPage': {
+    'ForeignDirectInvestmentFormSuccessPage': {
         'business_unit': 'Invest',
-        'site_section': 'HighPotentialOpportunities',
+        'site_section': 'ForeignDirectInvestment',
         'site_subsection': 'FormSuccessPage',
     },
     'InvestRegionPage': {